fixing paths for test_academy_csv_transform.py
